Fix an issue where only a single <polymer-signals> element operated per page.
In browsers without native Shadow DOM support, (like most Chrome releases now that the polyfill support has been switched on in Polymer), `<polymer-signals>` will not refire the same event. This has the effect of breaking every `<polymer-signals>` element on the page that is not the first one.

This PR simply generates a new event for each signaler to work around this issue.